added some more comments and refactored folder arrangement slightly
