+++ Freight/Assets/Scenes/TrainStationPun/NavMesh-NavMesh Surface.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 252bc56fca98d67489989eee1241e0d7
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
