--- Freight/Assets/Scripts/Photon/Lobby/Map.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 509100abfc6b8f34dbec2e063a844352
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
