--- Freight/Assets/Scripts/Photon/Lobby/Map/BagSpawner.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 96a7e610d2fad3b44906d4c47a4f5b2e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
