--- Freight/Assets/Scripts/Photon/Lobby/Map/CarWheelAnimation.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d3fcc7b91fcb90f41ab0b26c81514949
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
