--- Freight/Assets/Scripts/Photon/Lobby/Map/DropSpots.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a215522ad29aef64e8897ff0db2fa5e7
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
