--- Freight/Assets/Scripts/Photon/Lobby/Map/EndGame.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f7c599be652065141a7be2501b3bd8e0
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
