--- Freight/Assets/Scripts/Photon/Lobby/Map/GameSettings.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 4de07633da0909f4ab2cf609e1408bef
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
