--- Freight/Assets/Scripts/Photon/Lobby/Map/InitialText.cs
@@ -1,18 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class InitialText : MonoBehaviour
-{
-    // Start is called before the first frame update
-    void Start()
-    {
-        StartCoroutine(SelfDestruct());
-    }
-
-    IEnumerator SelfDestruct()
-    {
-        yield return new WaitForSeconds(3f);
-        Destroy(gameObject);
-    }
-}
