--- Freight/Assets/Scripts/Photon/Lobby/Map/InitialText.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2708442d5d7c1b44a88455b391337354
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
