--- Freight/Assets/Scripts/Photon/Lobby/Map/MapSpawnerPhoton.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 4b8254eed2ac8d440b94f8ab6b390639
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences:
-  - tree: {fileID: 1948757798113483429, guid: dd3e983fd184775429dd64a27bd326f5, type: 3}
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
