--- Freight/Assets/Scripts/Photon/Lobby/Map/Spotlights.cs
@@ -1,18 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using Photon.Pun;
-
-public class Spotlights : MonoBehaviour
-{
-    
-    // Start is called before the first frame update
-    void Start()
-    {
-        if (!(bool) PhotonNetwork.CurrentRoom.CustomProperties["SpotlightsActive"])
-        {
-            gameObject.SetActive(false);
-        }
-    }
-
-}
