--- Freight/Assets/Scripts/Photon/Lobby/Map/Spotlights.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 41ee8cb3d0b98994e96c2d16c6f9eb65
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
