--- Freight/Assets/Scripts/Photon/Lobby/Map/SyncedTime.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: c56a9b1599f76624396041ca3860d60b
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
