--- Freight/Assets/Scripts/Photon/Lobby/Map/SyncedTime/Freight.Scripts.Photon.Map.SyncedTime.SyncedTime.asmdef
@@ -1,16 +1,0 @@
-{
-    "name": "Freight.Scripts.Photon.Map.SyncedTyme",
-    "references": [
-        "GUID:57c32fc907df0f54e8e6e8f0d2488336",
-        "GUID:831409e8f9d13b5479a3baef9822ad34"
-    ],
-    "includePlatforms": [],
-    "excludePlatforms": [],
-    "allowUnsafeCode": false,
-    "overrideReferences": false,
-    "precompiledReferences": [],
-    "autoReferenced": true,
-    "defineConstraints": [],
-    "versionDefines": [],
-    "noEngineReferences": false
-}
