--- Freight/Assets/Scripts/Photon/Lobby/Map/SyncedTime/Freight.Scripts.Photon.Map.SyncedTime.SyncedTime.asmdef.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 063ed77af6700eb47b6a34fbd30c7201
-AssemblyDefinitionImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
