+++ Freight/Assets/Scripts/Photon/Map/Map.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: eabce54544ab8f542a13af4b267f9b7e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
