+++ Freight/Assets/Scripts/Photon/Map/Map/EndGameSecond.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2c10804cb3c10a248a4ea585dccb977e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
