+++ Freight/Assets/Scripts/Photon/Map/Map/SyncedTime/Freight.Scripts.Photon.Map.SyncedTime.SyncedTime.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 063ed77af6700eb47b6a34fbd30c7201
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
