+++ Freight/Assets/Scripts/Photon/Map/Map/SyncedTime/SyncedTime.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: dda4ff93fc7d77c4499c580e3c501711
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
