+++ Freight/Assets/Scripts/Photon/Player/Achievements.cs
@@ -34,7 +34,7 @@ public class Achievements : MonoBehaviourPun
 
     private Queue<IEnumerator> coroutineQueue = new Queue<IEnumerator>();
 
-    // this is a checker 
+    // this is a checker to see if the player has been detected by the guards at least once
     public bool wasDetectedOnce = false;
 
     // Start is called before the first frame update
